+++ docs/attribute-curator/CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this gear are documented in this file.
 
 ## 0.2.7 (unreleased)
 
+* Adds ability to blacklist NACCIDs from being curated
 * Adds the `ncrad_samples` scope
-* Adds a `curated` tag to each curated file, and will skip curation on subsequent passes if the file has this tag
+* Adds configuration to apply a tag to each curated file, and will skip curation on subsequent passes if the file has this tag
 
 ## 0.2.6